Move SearchForm components under data folder
