+++ react-components/src/data.js
@@ -1,3 +1,7 @@
 import * as details from "./data/details";
+import * as search from './data/search';
 
-export  {details};
+export {
+    details,
+    search
+};
