+++ react-components/src/data/search.js
@@ -1,7 +1,7 @@
 import SearchFormBase from './search/SearchFormBase';
 import SaveSearchButtonBase from './search/SaveSearchButtonBase';
 import {hasI18N} from '../util/hasI18N';
-import messages from '../diskResource/search/messages';
+import messages from './search/messages';
 import {withStoreProvider} from "../util/withStoreProvider";
 
 const SearchForm = withStoreProvider(hasI18N(SearchFormBase, messages));
