+++ react-components/src/data/search/SearchFormBase.js
@@ -7,9 +7,9 @@ import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
 import PropTypes from 'prop-types';
 import Button from '@material-ui/core/Button';
-import {SaveSearchButton} from '../../diskResource/search';
+import {SaveSearchButton} from '../search';
 import Select from '@material-ui/core/Select';
-import TagPanel from '../../data/details/TagPanel';
+import TagPanel from '../details/TagPanel';
 import TextField from '@material-ui/core/TextField';
 import injectSheet from 'react-jss';
 import styles from './styles';
