+++ react-components/src/data/search/styles.js
