--- react-components/src/diskResource.js
@@ -1,3 +1,0 @@
-import * as search from './diskResource/search'
-
-export { search };
