+++ react-components/src/lib.js
@@ -1,6 +1,5 @@
 import * as apps from "./apps";
 import * as data from "./data";
-import * as diskResource from "./diskResource"
 import React from "react";
 import ReactDOM from "react-dom";
 import {MuiThemeProvider, createMuiTheme} from "@material-ui/core/styles"; // v1.x
@@ -16,7 +15,6 @@ const getDefaultTheme = () => createMuiTheme({});
 export {
     apps,
     data,
-    diskResource,
     getDefaultTheme,
     MuiThemeProvider
 };
