+++ react-components/stories/index.stories.js
@@ -10,23 +10,22 @@ import TagTest from "./data/details/Tag.stories";
 import TagPanelTest from "./data/details/TagPanel.stories";
 import AppStatsTest from "./apps/admin/AppStats.stories";
 import DEHyperLinkTest from "./util/hyperlink/DEHyperLink.stories";
-import SaveSearchButtonTest from './diskResource/search/SaveSearchButton.stories'
-import SearchFormTest from "./diskResource/search/SearchForm.stories";
+import SaveSearchButtonTest from './data/search/SaveSearchButton.stories'
+import SearchFormTest from "./data/search/SearchForm.stories";
 
+storiesOf('apps/admin/AppStats', module).add('with test stats', () => <AppStatsTest/>);
 storiesOf('apps/details', module).add('CategoryTree', () => <CategoryTreeTest logger={action('hierarchy')} />);
 storiesOf('apps/details', module).add('ToolDetails', () => <ToolDetailsTest/>);
-storiesOf('apps/admin/AppStats', module).add('with test stats', () => <AppStatsTest/>);
 
 storiesOf('data/BasicDetails', module).add('with test diskresource details', () => <BasicDetailsTest logger={action('details')}/>);
 storiesOf('data/InfoTypeSelectionList',module).add('with test diskresource details', () =>  <InfoTypeSelectionListTest logger={action('infoTypes')}/>);
+storiesOf('data/search', module).add('SearchForm', () => <SearchFormTest searchLogger={action('Search Submitted')}
+                                                                         saveSearchLogger={action('Save Search')}
+                                                                         editTagLogger={action('Edit Tag Selected')}
+                                                                         suggestionsLogger={action('Fetch Suggestions')}/>);
+storiesOf('data/search', module).add('SaveSearchButton', () => <SaveSearchButtonTest logger={action('Save Search')}/>);
 storiesOf('data/Tag', module).add('with test diskresource details', () => <TagTest logger={action('tag')}/>);
-storiesOf('data/TagPanel', module).add('witht test diskresource details', () => <TagPanelTest logger={action('tagpanel')}/>);
-
-storiesOf('diskResource/search', module).add('SearchForm', () => <SearchFormTest searchLogger={action('Search Submitted')}
-                                                                                 saveSearchLogger={action('Save Search')}
-                                                                                 editTagLogger={action('Edit Tag Selected')}
-                                                                                 suggestionsLogger={action('Fetch Suggestions')}/>);
-storiesOf('diskResource/search', module).add('SaveSearchButton', () => <SaveSearchButtonTest logger={action('Save Search')}/>);
+storiesOf('data/TagPanel', module).add('with test diskresource details', () => <TagPanelTest logger={action('tagpanel')}/>);
 
 storiesOf('util', module).add('CopyTextArea', () => <CopyTextAreaTest/>);
 storiesOf('util', module).add('DEHyperLink', () => <DEHyperLinkTest/>);
